reorganized some parts. end of day commit.
