+++ src/version.py
@@ -1,6 +1,9 @@
 # version.py
 # contains version information
 
+# global todos go here:
+# TODO: figure out what relay 2 does
+
 __version_info = (0, 1, 0)
 __version__ = '.'.join(map(str, __version_info))
